refactor: Move internal/cli to cli/
The goal of this commit is to simplify the package hierarchy a little
bit. Using internal/ unnecessarily complicates the package hierarchy in
this specific case. Most things are not exported anyways, and it adds an
extra indirection for exporting from the cli package.
