+++ cli/hooks.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package cmd
+package cli
 
 import (
 	"github.com/spf13/pflag"
